split cli shared recovery tests.
