+++ cli/tests/integration/shared_recovery/mod.rs
@@ -1,0 +1,4 @@
+mod create;
+mod delete;
+mod info;
+mod list;
